# docs/swagger.yaml
definitions:
  models.Account:
    properties:
      account_id:
        type: integer
      balance:
        type: number
    type: object
  models.AccountCreateRequest:
    properties:
      account_id:
        type: integer
      initial_balance:
        type: string
    type: object
  models.TransactionRequest:
    properties:
      amount:
        type: string
      destination_account_id:
        type: integer
      source_account_id:
        type: integer
    type: object
info:
  contact: {}
paths:
  /accounts:
    post:
      consumes:
      - application/json
      description: Create a new account
      parameters:
      - description: Account create request
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.AccountCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create account
      tags:
      - accounts
  /accounts/{account_id}:
    get:
      description: Get account by ID
      parameters:
      - description: Account ID
        in: path
        name: account_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Account'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get account
      tags:
      - accounts
  /transactions:
    post:
      consumes:
      - application/json
      description: Create a new transaction
      parameters:
      - description: Transaction request
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/models.TransactionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TransactionRequest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create transaction
      tags:
      - transactions
swagger: "2.0"