# internal/models/transaction.go
package models

// TransactionRequest represents a transfer between accounts
type TransactionRequest struct {
	SourceAccountID      int64  `json:"source_account_id"`
	DestinationAccountID int64  `json:"destination_account_id"`
	Amount               string `json:"amount"`
}

// Transaction represents a database transaction record
type Transaction struct {
	ID                   int64  `json:"id"`
	SourceAccountID      int64  `json:"source_account_id"`
	DestinationAccountID int64  `json:"destination_account_id"`
	Amount               string `json:"amount"`
}